# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Airstory API
  version: 1.0.0
  description: >
    REST API for the Airstory web app.
    Firebase is the identity provider. Protected endpoints require a Firebase ID token as a
    Bearer credential; sign-in, token refresh, sign-out, and password changes happen client-side
    via the Firebase SDK. See docs/openapi.yaml (this file) for the full contract.
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://air-sensor-api.onrender.com/api
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: >
        Firebase ID token from the client SDK. The backend verifies it with Firebase Admin
        and maps the Firebase uid to the app user via users.firebase_uid.
  schemas:
...
    StatSummary:
      type: object
      properties:
        mean:
          type: number
          nullable: true
        min:
          type: number
          nullable: true
        max:
          type: number
          nullable: true
        median:
          type: number
          nullable: true
        stddev:
          type: number
          nullable: true
        sample_count:
          type: integer
    HeatmapPoint:
      type: object
      properties:
        latitude:
          type: number
          description: Rounded to 4 decimal places
        longitude:
          type: number
          description: Rounded to 4 decimal places
        value:
          type: number
          description: Average metric value at this location
        point_count:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
paths:
...
  /workspaces/{workspaceId}/analytics/summary:
    get:
      tags: [Analytics]
      summary: Get aggregate statistics for a metric in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: metric
          in: query
          schema:
            type: string
            enum: [pm25, co, temp, humidity]
            default: pm25
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  metric:
                    type: string
                  summary:
                    $ref: '#/components/schemas/StatSummary'
        '400':
          description: Invalid metric
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/heatmap:
    get:
      tags: [Analytics]
      summary: Get aggregated heatmap points for a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: metric
          in: query
          schema:
            type: string
            enum: [pm25, co, temp, humidity]
            default: pm25
      responses:
        '200':
          description: Heatmap data points
          content:
            application/json:
              schema:
                type: object
                properties:
                  metric:
                    type: string
                  points:
                    type: array
                    items:
                      $ref: '#/components/schemas/HeatmapPoint'
        '400':
          description: Invalid metric
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/export/measurements.csv:
    get:
      tags: [Analytics]
      summary: Export workspace measurements as a CSV file
      description: Returns up to 5000 measurements, most recent first.
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CSV file download
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename=measurements.csv
          content:
            text/csv:
              schema:
                type: string
                description: >
                  Columns: id, captured_at, pm25, co, temp, humidity,
                  latitude, longitude, indoor_outdoor, session_code,
                  session_name, school_code, instructor, period, group_code
  /analytics/openaq/daily:
    get:
      tags: [Analytics]
      summary: Fetch daily reference air quality data from OpenAQ
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2024-01-01'
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2024-01-31'
        - name: metric
          in: query
          schema:
            type: string
            default: pm25
      responses:
        '200':
          description: OpenAQ daily measurements (variable structure)
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid coordinates or missing dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: OpenAQ API key not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /analytics/openaq/heatmap:
    get:
      tags: [Analytics]
      summary: Fetch nearby OpenAQ sensor locations for a heatmap overlay
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: metric
          in: query
          schema:
            type: string
            default: pm25
        - name: radius
          in: query
          schema:
            type: number
            default: 15000
            description: Search radius in meters
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
            description: Maximum number of sensor locations to return
      responses:
        '200':
          description: OpenAQ sensor locations (variable structure)
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: OpenAQ API key not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Sheets ────────────────────────────────────────────────────────────────
...